Rename rand::distributions -> rand::distr
