+++ src/rng.rs
@@ -9,8 +9,8 @@
 
 //! [`Rng`] trait
 
-use crate::distributions::uniform::{SampleRange, SampleUniform};
-use crate::distributions::{self, Distribution, Standard};
+use crate::distr::uniform::{SampleRange, SampleUniform};
+use crate::distr::{self, Distribution, Standard};
 use core::num::Wrapping;
 use core::{mem, slice};
 use rand_core::RngCore;
@@ -86,7 +86,7 @@ pub trait Rng: RngCore {
     /// rng.fill(&mut arr2);                    // array fill
     /// ```
     ///
-    /// [`Standard`]: distributions::Standard
+    /// [`Standard`]: distr::Standard
     #[inline]
     fn random<T>(&mut self) -> T
     where
@@ -125,7 +125,7 @@ pub trait Rng: RngCore {
     /// println!("{}", n);
     /// ```
     ///
-    /// [`Uniform`]: distributions::uniform::Uniform
+    /// [`Uniform`]: distr::uniform::Uniform
     #[track_caller]
     fn gen_range<T, R>(&mut self, range: R) -> T
     where
@@ -139,7 +139,7 @@ pub trait Rng: RngCore {
     /// Generate values via an iterator
     ///
     /// This is a just a wrapper over [`Rng::sample_iter`] using
-    /// [`distributions::Standard`].
+    /// [`distr::Standard`].
     ///
     /// Note: this method consumes its argument. Use
     /// `(&mut rng).gen_iter()` to avoid consuming the RNG.
@@ -154,7 +154,7 @@ pub trait Rng: RngCore {
     /// assert_eq!(&v, &[1, 2, 3, 4, 5]);
     /// ```
     #[inline]
-    fn gen_iter<T>(self) -> distributions::DistIter<Standard, Self, T>
+    fn gen_iter<T>(self) -> distr::DistIter<Standard, Self, T>
     where
         Self: Sized,
         Standard: Distribution<T>,
@@ -168,7 +168,7 @@ pub trait Rng: RngCore {
     ///
     /// ```
     /// use rand::{thread_rng, Rng};
-    /// use rand::distributions::Uniform;
+    /// use rand::distr::Uniform;
     ///
     /// let mut rng = thread_rng();
     /// let x = rng.sample(Uniform::new(10u32, 15).unwrap());
@@ -189,7 +189,7 @@ pub trait Rng: RngCore {
     ///
     /// ```
     /// use rand::{thread_rng, Rng};
-    /// use rand::distributions::{Alphanumeric, Uniform, Standard};
+    /// use rand::distr::{Alphanumeric, Uniform, Standard};
     ///
     /// let mut rng = thread_rng();
     ///
@@ -213,7 +213,7 @@ pub trait Rng: RngCore {
     ///     println!("Not a 6; rolling again!");
     /// }
     /// ```
-    fn sample_iter<T, D>(self, distr: D) -> distributions::DistIter<D, Self, T>
+    fn sample_iter<T, D>(self, distr: D) -> distr::DistIter<D, Self, T>
     where
         D: Distribution<T>,
         Self: Sized,
@@ -259,11 +259,11 @@ pub trait Rng: RngCore {
     ///
     /// If `p < 0` or `p > 1`.
     ///
-    /// [`Bernoulli`]: distributions::Bernoulli
+    /// [`Bernoulli`]: distr::Bernoulli
     #[inline]
     #[track_caller]
     fn gen_bool(&mut self, p: f64) -> bool {
-        match distributions::Bernoulli::new(p) {
+        match distr::Bernoulli::new(p) {
             Ok(d) => self.sample(d),
             Err(_) => panic!("p={:?} is outside range [0.0, 1.0]", p),
         }
@@ -291,11 +291,11 @@ pub trait Rng: RngCore {
     /// println!("{}", rng.gen_ratio(2, 3));
     /// ```
     ///
-    /// [`Bernoulli`]: distributions::Bernoulli
+    /// [`Bernoulli`]: distr::Bernoulli
     #[inline]
     #[track_caller]
     fn gen_ratio(&mut self, numerator: u32, denominator: u32) -> bool {
-        match distributions::Bernoulli::from_ratio(numerator, denominator) {
+        match distr::Bernoulli::from_ratio(numerator, denominator) {
             Ok(d) => self.sample(d),
             Err(_) => panic!(
                 "p={}/{} is outside range [0.0, 1.0]",
@@ -554,7 +554,7 @@ mod test {
 
     #[test]
     fn test_rng_trait_object() {
-        use crate::distributions::{Distribution, Standard};
+        use crate::distr::{Distribution, Standard};
         let mut rng = rng(109);
         let mut r = &mut rng as &mut dyn RngCore;
         r.next_u32();
@@ -566,7 +566,7 @@ mod test {
     #[test]
     #[cfg(feature = "alloc")]
     fn test_rng_boxed_trait() {
-        use crate::distributions::{Distribution, Standard};
+        use crate::distr::{Distribution, Standard};
         let rng = rng(110);
         let mut r = Box::new(rng) as Box<dyn RngCore>;
         r.next_u32();
